Time: 1189 ms (76.53%), Space: 154.8 MB (45.10%) - LeetHub

+++ 0446-arithmetic-slices-ii-subsequence/0446-arithmetic-slices-ii-subsequence.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     int numberOfArithmeticSlices(vector<int>& nums) {
-        long long sum = 0;
-        vector<unordered_map<long long, long long>> dp(nums.size());
-        for(long long i = 0; i < nums.size(); i++){
-            for(long long j = 0; j < i; j++){
-                long long diff = (long long)nums[i] - (long long)nums[j];
+        int sum = 0;
+        vector<unordered_map<long,long>> dp(nums.size());
+        for(int i = 0; i < nums.size(); i++){
+            for(int j = 0; j < i; j++){
+                long diff = (long)nums[i] - (long)nums[j];
                 dp[i][diff] += 1;
                 if(dp[j].count(diff) != 0){
                     sum += dp[j][diff];
